Create submission dir if not present
